requirements.txt
pyautogen
gradio
groq
playwright
beautifulsoup4
anthropic